rename 'experiment' to 'menu/index'
