+++ src/pages/c/menu.astro
@@ -2,7 +2,7 @@
 import Layout from '../../layouts/Layout.astro'
 
 // test data
-import { Menu as AMenu, type MenuItem } from '../../components/table/menu/experiment'
+import { Menu as AMenu, type MenuItem } from '../../components/table/menu'
 
 const menuItems: MenuItem[] = [
   { value: 'Home', label: 'Home' },
